mas_execution_manager/requirements.txt
git+https://github.com/ropod-project/ftsm.git
kafka-python
jsonschema
pymongo
numpy
PyYAML
